Issue-3499: Move constants module to compose/common package
 In order to figure decouple CLI from Core it is necessary
 to understand their cross-dependencies and decouple them
 for further compose/core extraction
